requirements.txt
Django>=5.2.10
djangorestframework>=3.16.1
psycopg2-binary>=2.9.11
celery>=5.6.2
redis>=7.1.0
requests>=2.32.5
flower>=2.0